requirements.txt
streamlit
pandas
openai-whisper
openai
audio-recorder-streamlit